+++ services/web/src/API.hs
@@ -153,6 +153,7 @@ import API.Dashboard.Users.Detail.Get.Handler qualified as Dashboard.Users.Detai
 import API.Dashboard.Users.Edit.Get.Handler qualified as Dashboard.Users.Edit.Get
 import API.Dashboard.Users.Edit.Post.Handler qualified as Dashboard.Users.Edit.Post
 import API.Dashboard.Users.Get.Handler qualified as Dashboard.Users.Get
+import API.Dashboard.Users.ResetPassword.Post.Handler qualified as Dashboard.Users.ResetPassword.Post
 import API.Dashboard.Users.Role.Patch.Handler qualified as Dashboard.Users.Role.Patch
 import API.Dashboard.Users.Suspend.Post.Handler qualified as Dashboard.Users.Suspend.Post
 import API.Dashboard.Users.Unsuspend.Post.Handler qualified as Dashboard.Users.Unsuspend.Post
@@ -543,6 +544,7 @@ server =
           rolePatch = Dashboard.Users.Role.Patch.handler,
           suspendPost = Dashboard.Users.Suspend.Post.handler,
           unsuspendPost = Dashboard.Users.Unsuspend.Post.handler,
+          resetPasswordPost = Dashboard.Users.ResetPassword.Post.handler,
           delete = Dashboard.Users.Delete.handler
         }